+++ backend.py
@@ -2,6 +2,8 @@ import sys
 from configparser import ConfigParser
 from chatbot import ChatBot
 import webbrowser
+from selenium import webdriver
+from bs4 import BeautifulSoup
 
 def generate_html(output_text):
     # Replace newline characters with <br> for line breaks
@@ -53,7 +55,26 @@ def generate_html(output_text):
     with open("chatbot_response.html", "w", encoding="utf-8") as html_file:
         html_file.write(html_content)
 
+def scrape_content_from_url(url):
+    driver = webdriver.Edge()  
+    driver.get(url)
+
+    soup = BeautifulSoup(driver.page_source, 'html.parser')
+    visible_text = soup.get_text()
+
+    driver.quit()
+    return visible_text
+
 def main():
+    url = "file:///C:/Users/kushp/Downloads/Dark%20Pattern%20Hackathon%20(3).pdf"
+
+    content = scrape_content_from_url(url)
+
+    with open('output.txt', 'w', encoding='utf-8') as file:
+        file.write(content)
+
+    print(f'Content extracted from the page and saved to "output.txt".')
+
     try:
         # Read API key from credentials.ini
         config = ConfigParser()
