+++ extension/manifest.json
