+++ extension/popup.js
