+++ extension/window.js
